backend-files/gateways/goodwe-sems/requirements.txt
# requirements.txt
fastapi
uvicorn
requests
python-dotenv
